Feat: 예외 케이스 Test 파일 분리
What is this PR?
- ErrorTest 파일 작성
